Improved error handling.
It could still happen that we do not consume Z ReadyForQuery after an error.
We introduce a new connection state to detect this.

Also, previously it was possible that DDL may succeed on the coordinator
and get committed but not on the datanodes. We now make sure it does
not get committed on the coordinator.