Modified pgbench for Postgres-XC.
By default, it will use branch id to distribute by, and also include it in the queries.
This allows for better multi-node scalability.  For this configuration, we do not
have to worry about implicit two phase commits occurring.

If running in either initialization or execution mode, alternatively a new option
-k may be used (-k = "key" = no branch id).

If included with -i, tables are distributed according to their primary key.
This means we will see a lot of two phase commit activity over multiple data nodes.

Similarly, if -k is included when running tests, branchid (bid) will NOT
be used as part of the DML statements. This helps us measure the impact of 2PC.