This is the first of some planned changes to recognize more "Postgres-XC safe" queries, before we implement more complex cross-node operations.
This focuses on native pg_catalog based views like pg_settings.
This is because it used by pgadmin. With the attached patch, pgadmin no
longer complains about such queries because they are now supported.

The PostgreSQL query rewriter appears to convert the view into a subquery.
I modified the XC planner to check for pg_catalog table usage in the
FROM clause (and subqueries thereof).

In addition, pgadmin was sending "SET client_encoding to 'UNICODE'".
It was being swallowed and proxied by Postgres-XC, but no response was sent
back because it did not recognize the 'S' message response. I have now
added support for that, too.

Note that we have planned a whole separate task planned for proper SET
handling. These commands will now be processed, but the user should not
assume that these will remain set, due to Postgres-XC pooler.
Again, we will address this in the future.

With this patch pgadmin runs much more smoothly when connected to Postgres-XC.
Users can view schema information and issue SQL statements.  There is
still an error window that pops up when clicking on a table because it
issues a "SELECT count(*) FROM table", which is not yet supported in
Postgres-XC. This problem will disappear soon however, as a patch is
almost ready for basic aggregate support.