There is a race condition that could lead to problems for the CLOG and sub transactions.
In Postgres-XC, multiple processes may decide to extend the CLOG
at the same time. One will wait for the other, then afterwards
re-zero out the page.

Instead, once the lock is obtained, we re-check to make sure
that another process did not already extend and create the page.
If so, we just exit.