Improved error handling.
The primary focus is to better handle the case of a stopped or
crashed data node on the coordinator.

Also, before a rollback make sure connections are clean.

If there was an error, tell the pooler to destroy the connections instead
of returning them to the pools, even the data node connections that did
not have an error but are involved in the statement. This is becaue there
may be some remaining messages buffered or in transit, and could affect
subsequent requests.