Support for CLEAN CONNECTION
Utility to clean up Postgres-XC Pooler connections.
This utility is launched to all the Coordinators of the cluster

Use of CLEAN CONNECTION is limited to a super user.
It is advised to clean connections before dropping a Database.

SQL query synopsis is as follows:
CLEAN CONNECTION TO
             (COORDINATOR num | DATANODE num | ALL {FORCE})
              FOR DATABASE dbname
Connection cleaning has to be made on a chosen database called dbname.

It is also possible to clean connections of several Coordinators or Datanodes
Ex:  CLEAN CONNECTION TO DATANODE 1,5,7 FOR DATABASE template1
             CLEAN CONNECTION TO COORDINATOR 2,4,6 FOR DATABASE template1

Or even to all Coordinators/Datanodes at the same time
Ex:  CLEAN CONNECTION TO DATANODE * FOR DATABASE template1
             CLEAN CONNECTION TO COORDINATOR * FOR DATABASE template1

When FORCE is used, all the transactions using pooler connections are aborted,
and pooler connections are cleaned up.
Ex:  CLEAN CONNECTION TO ALL FORCE FOR DATABASE template1;

FORCE can only be used with TO ALL, as it takes a lock on pooler to stop requests
asking for connections, aborts all the connections in the cluster, and cleans up
pool connections