Add support for INSERT SELECT. Also, initial changes for setting execution nodes in general planner.
We execute a query normally, directed by the Coordinator,
and then use COPY running down on the data nodes to insert
the data there.

We also check for a special case of INSERT SELECT where the source
and destination columns are both the distribution columns of
tables, and where the query is a single step query. In such a
case, we just execute the query locally on the data nodes without
the use of COPY directed by the Coordinator.

In testing we uncovered an issue for INSERT SELECT when the
query is from a replicated table; it was selecting from all
tables because exec_nodes was not set for the RemoteQuery.
As a result, this commit also sets the exec_nodes for base
RemoteQuery structs as well as join reduced ones.

It does not yet, however, take into account WHERE clause
equality conditions against the distribution column, as
is done in a regular SELECT.  This is left as a future
optimization, best done as a step for further merging
the Postgres-XC planner and the standard planner.