Correction for implicit 2PC
When a COMMIT is issued for a write transaction involving multiple
Postgres-XC nodes, a 2PC is used internally.
For implicit 2PC, the following process is respected for DDL transactions:
1) PREPARE on local Coordinator (if DDL is involved)
2) PREPARE on Postgres-XC nodes
3) COMMIT PREPARED on local Coordinator (if DDL is involved)
4) COMMIT PREPARED on Postgres-XC nodes

For transaction containing no DDL:
1) PREPARE on Datanodes
2) COMMIT on Coordinator
3) COMMIT PREPARED on Datanodes

In case of a Node failure after Coordinator has committed,
transaction becomes partially committed on Nodes.
To maintain data consistency, it is absolutely necessary to COMMIT this transaction on all nodes.
In this case, the remaining list PREPARED nodes is saved on GTM as if it was an explicit 2PC.
And this transaction is kept open to avoid visibility issues.

It is necessary to issue a COMMIT PREPARED from application to finish the COMMIT
of this transaction.