Updated regression tests.
In Postgres-XC, we cannot be sure of the return order of results, so
the tests were updated with ORDER BY added to many SELECT statements.

There are still many tests that fail, but this at least gives us a
better baseline to more easily find and fix the open bugs.

Written by Benny