Node Registering feature
Feature to register a Proxy, Coordinator or Datanode in Postgres-XC cluster.
When a PGXC Node connects in the cluster,
Node connects to GTM to register the following information:
- Node ID
- Node Type
- Data folder
- Port number
- Socket IP (got directly from remote socket when registering)
- Proxy Number Node goes through (0 if no Proxy used)

When a node is shutdown, it automatically unregisters itself.
If node is quickly shutdown, it disconnects itself, and can reconnect itself later.
In this case, node ID is reserved for this node.

When a node tries to register with an ID already used, it cannot start up.
Socket IP is forwarded to GTM if Node goes through a Proxy.

As proxy also registers, a start-up option "-i" has been added to set Proxy ID.