The patch implements multiple insert syntax in PGXC.
Multiple insert means using a single insert statement to insert
multiple rows into a table using the syntax e.g.
insert into students(rno, class, pos) values
(1, 10, 5), (2, 10, 6), (3, 10, 7), (4, 10, 8);

Without the patch statements like these pass,
but actually do not insert any thing in the table.

The main code changes are in re-writer.
The patch checks to see if the insert statement
has more than one sets in the provided list of values
(FOUR in the above example), and in that case rewrites the insert statement.
The insert rewriter separates the sets in the provided
list of values into independent lists depending on the
distribution of the table, the distribution column and
the value provided for the distribution column.

Next the main re-writer is separated into two possible
paths, one without a for loop and if we have a separated
list of insert values, we run a for loop on the list
and create an insert statement for each of the data nodes
providing it that sub-group of the original list
that is supposed to run on this particular data node.

Main work is done now, all that is left is to handle
multiple command result tags from the data nodes.
HandleCmdComplete does this, it simply keeps adding
into the insert row count until all data nodes are done.

With this patch, multi insert does not work for replicated tables.
Additional comments are also necessary.