CREATE TABLE: new distribution function MODULO
new catalogue is named pgxc_class.
It has the following columns

pcrelid       - The OID of the added relation in pg_class
pclocatortype - XC defines distribution types as a single
                character in locator.h
                The provided distribution type is stored here.
                Distribution types can be
        'R' - Replicated
        'H' - Based on a column's hash value
        'N' - Based on round robin for all data nodes
        'M' - Based on modulo of a column's value
              This new distribution type is added by this patch.
        Similarly G, S and C are also defined.
pcattnum      - The number of the column used for replication
        valid only for H and M - Zero for the rest

This patch supports also copy and INSERT(SELECT).
It does NOT SUPPPORT multiple INSERT.

The following issues are remaining though:
1) Non-definition of distribution column in table makes a crash
2) Creation of an index on a column that is not the distribution column
3) Creation of a foreign key on distribution column
4) Distribution on another column other than unique makes a crash