Fix for bug 3134395, 3086422, 3136230: 2PC locks
When a PREPARE was being made on Coordinator for a transaction not
using DDL, 2PC file was bypassed so as to make a fake COMMIT on Coordinator.
The problem was that locks hold at PREPARE state were not released at COMMIT
PREPARED on Coordinator because lock information was obtained from 2PC,
which indeed did not exist on Coordinator for non-DDL transactions.

With this fix, instead of a fake PREPARE, XC does a Commit on Coordinator
without contacting GTM, this permits to release correctly all the locks
held by transaction.

PrepareTransaction is simplified in xact.c, but CommitTransaction
needs an additional parameter to decide if GTM is called at COMMIT
or not.