This patch fixes a test case in strings.sql
SELECT CAST(f1 as text) AS "f2" FROM CHAR_TBL ORDER BY f1
was reporting

ERROR: ORDER BY "f1" is ambiguous

The problem was in the function reconstruct_step_query.
The function was reconstructing the query to look like

SELECT (f1)::text, f1 FROM CHAR_TBL ORDER BY f1

and when this query was being sent to the data node, it was complaining.

The solution adopted in this patch is to add column alias (if any) while reconstructing the query.

Some cosmetic changes were also required for the strings test to pass.