Performance issue with snapshot processing
This commit solves an issue with snapshot processing in the case
where cluster was used under multiple Coordinators.

When a Coordinator/Coordinator connection was initialized,
backend coordinator initialized a transaction to GTM that was never
committed. A consequence of that was Snapshot xmin that remained
to a constant value. Another consequence was recent global Xmin
value set to a lower value, making autovacuum having absolutely no effects.

This fix improves performance of the whole cluster by making autovacuum
remove correctly old tuples for long-period runs.