Reactivate preferred_data_nodes
preferred_data_nodes is a GUC parameter that allows the user to
set a list of node numbers in a string to determine on which node
a read operation should be done for replicated tables.
It happened that even if parameter was set, the list of preferred nodes
was not created from the GUC string and so it had no effects.

It has been found that this fix improves performance of the whole cluster
system by up to 30%.