Fix for bug 3438569: Prepared statement on distributed table
A prepared statement on a query using distributed tables targetted
always the first node when it should have targetted a node depending
on the expression value and expression type used to evaulate the list
of nodes at execution.
The expression value was correctly obtained but expression type
was not taken into account, leading to an error on node index
determination.

Conflicts:

	src/backend/pgxc/pool/execRemote.c