Fix for queries containing reducable-joins along with local quals, such as this: select ti1.col2 FROM tab1 ti1 left join tab2 ti2 on ti1.col1 = ti2.col1 where ti1.col2 = f(2); -- where f() is not immutable.
in create_remotejoin_plan(), we convert two joined RTEs into one single RTE,
but we do not change the target list and local quals to set their references
to this changed RTE, although we add this new RTE into the rte array. So, when
there are local quals, these local quals will not work with the new RTEs unless
we change the varnos and varattnos to those of the new RTE index.

Fix involved doing something like what set_plan_references() does at the end of
planning, i.e. setting varnos and varattnos to the new RTE index.
The field RemoteQuery->base_tlist is used to indicate that we need to do such
varno adjustments. This field is set by create_remotejoin_plan(), but
not create_remotegrouping_plan() because the later does not require adjustments to
the varno and varattnos.