Add support for INSERT INTO statement. We create a RemoteQuery plan for every resultRelationInfo and run the remote inserts instead of local inserts.
Add support for CREATE TABLE AS and SELECT INTO statements.
These queries are rewritten as CREATE TABLE followed by the INSERT INTO
statement.

Also fix issues with ROUND ROBIN distribution and PREPARE/EXECUTE. This was
an old bug, but became more evident after this work.

The old handling of multiple VALUES for INSERT statement is also now handled
through standard_planner. The old handling was causing segfaults with
PREPARE/EXECUTE

This fix should not have been committed to the stable branch. But since the
testing team needs a fix for the PREPARE/EXECUTE issue with multiple VALUES
clause, we decided to backport this.