Fix get_from_clause_item's logic to print alias in case of sub-query.
The code to print alias in case of sub query was added by the following commit
6dcef38adc1cad727f04dc18d1246519910ba767. The commit fixes the bug 3446932.
This fix was necessary because PG does not need to deparse a rewritten query and hence
does not need to take care of sub-query case while deparsing.

When this fix was done the name of the variable was gavealias, which has now become printalias
and its sense has also been flipped due to the following commit
8b728e5c6e0ce6b6d6f54b92b390f14aa1aca6db.

We therefore need to remove the code to add alias into query buffer to avoid printing
the alias twice.