Fix for join regression.
Please note that join regression still fails with this commit.

The failure happens at the statement with combination of outer join and
lateral.  The story is as follows:

In the vanilla PG, this query falls into parameterized plan.   Because
XC does not use statistics, this falls into usual plan, where this combination
is not handled correctly.  The fix is already available in further PG commit
in the master and REL_9_3_STABLE branch.   The core team decided to leave this
as is and test in the further merge with PG master and 9.3 lines.

Appreciate for detailed analysis by Ashutosh Bapat.