Fix for a bug while updating a view, taken from updatable_views.sql
CREATE TABLE base_tbl (a int PRIMARY KEY, b text DEFAULT 'Unspecified') distribute by replication;
INSERT INTO base_tbl SELECT i, 'Row ' || i FROM generate_series(-2, 2) g(i);

CREATE VIEW rw_view1 AS SELECT b AS bb, a AS aa FROM base_tbl WHERE a>0;
CREATE VIEW rw_view2 AS SELECT aa AS aaa, bb AS bbb FROM rw_view1 WHERE aa<10;

INSERT INTO rw_view2 VALUES (3, 'Row 3');
INSERT INTO rw_view2 (aaa) VALUES (4);
UPDATE rw_view2 SET bbb='Row 4' WHERE aaa=4;
ERROR:  cache lookup failed for attribute aaa of relation...

This test case was failing because of the following logic that was being
used to find the base table columns that are being updated.
In the rewritten query tree, iterate over the non-junk target list entries
and try to find the attribute number of the TargetEntry::resname in the target
table being updated. In case of failure report error
"cache lookup failed for attribute..."

The problem was that for the case of the example quoted above resname had
"aaa" which is not the name of any column in target table.

It was tried to improve the logic to find the columns being updated
while creating the remote update statement, but it made the code complex.
The code complexity arises for the case of a table with child tables,
where columns are added to parent using ALTER TABLE. The attribute
number of the added column is different in parnet and child table.
In this case we first have to use TargetEntry::resno to find the name
of the column being updated in parent table, and then find the attribute
number of that particular column in the child. This makes code complex.
In comaprison if we choose to update all the columns of the table
irrespective of the columns being updated, the code becomes simple
and easy to read.
Performance comparison between the two approaches (updating all columns
and updating only the columns that were in the target list) shows that
both the approaches give similar TPS in hour long runs of DBT1.