Fix for alias problem in returning list of INSERT statements
When the query deparsing system generates unique names for
rtable entries it changes the rtable name of the result relation by
appending digits to it. This patch changes the renaming scheme such
that the result relation name is kept as is whereas the names of the
other duplicate entries of the same table in rtable list are changed.
This is done to make sure that returning list entries in an INSERT
statement are not generated with changed name because Insert queries
do not support table aliases.