Fix false report of partition column updation.
Consider this test case taken from updatable_views.sql

CREATE TABLE base_tbl (a int, b int);
INSERT INTO base_tbl VALUES (1,2), (4,5), (3,-3);
CREATE VIEW rw_view1 AS SELECT * FROM base_tbl ORDER BY a+b;
EXPLAIN (verbose, costs off) UPDATE rw_view1 SET b = b + 1 RETURNING *;
results in
ERROR:  Partition column can't be updated in current version

The problem is that the rewritten query has column 'a' in the
target list and it is not detectable in validate_part_col_updatable
that the added column is because of query rewrite and is not being changed.

The fix is to detect that the update query has a construct dist_col = dist_col
in the set clause and in this case allow the update.

The fix was suggested and the patch was reviewed by Ashutosh Bapat.