This is a fix for triggers regression.   COPY commands are replaced with INSERT to fire row triggers (XC now dows not support row triggers with COPY commands).
The regression has some new use of temporary objects so
enforce_two_phase_commit has been turned off.

Please note that the regression still fails.  The regression has new
tests for the fix of the bug #6123.  It was found even in PostgreSQL
9.2 and was fixed after then.  Unfortunately, XC needs some more work
to fix it.   Fortunately, the bug is not easy to reproduce and no
report has been made from the field.